refactor(server): extract runtime helpers from main.rs into runtime.rs
main.rs had grown to 7917 lines (fc-quality-audit 2026-06-13 flagged it
CRITICAL). Move the self-contained tail — process lifecycle (panic hook,
fatal_exit, shutdown_signal), the container healthcheck probe, listener
binding + port handshake, the basic-auth parser, and the
POST /_fakecloud/wafv2/evaluate admin handler — plus their unit tests into
a new runtime.rs module. All are parameterized free functions that capture
no main() locals, so this is a behaviour-preserving move; main() imports
them back and every call site is unchanged.

main.rs: 7917 -> 7378 lines. Establishes the sibling-module pattern for
further decomposition of the router/state-init body.

Found by fc-quality-audit 2026-06-13.
